+++ src/middlewares/attackCurrentUser.js
@@ -8,11 +8,10 @@ module.exports = async (req, res, next) => {
 
     const user = await UserModel.findOne(
       { where: { id: loggedInUser.id } },
-      { passwordHash: 0, __v: 0 } // Excluindo o hash da senha da resposta que vai pro servidor, por segurança
+      { passwordHash: 0, __v: 0 }
     );
 
     if (!user) {
-      // 400 significa Bad Request
       return res.status(400).json({ msg: "User does not exist." });
     }
