+++ src/middlewares/isAuthenticated.js
@@ -1,6 +1,6 @@
 const jwt = require("express-jwt");
 
-function extractTokenFromHeaders(req, res) {
+function extractTokenFromHeaders(req) {
   if (!req.headers.authorization) {
     // O throw tem o mesmo efeito do return de encerrar a execução da função
     throw new Error("Missing Authorization Header.");
