+++ src/models/client.model.js
@@ -14,12 +14,27 @@ const Client = db.define("client", {
     type: DataTypes.STRING,
     allowNull: false,
     validate: {
-      notEmpty: true,
+      notEmpty: {
+        msg: "The field cannot be empty!",
+      },
+      len: {
+        args: [2, 20],
+        msg: "This field must be between 4 and 20 characters!",
+      },
     },
   },
   cpf: {
     type: DataTypes.STRING,
     unique: true,
+    validate: {
+      notEmpty: {
+        msg: "The field cannot be empty!",
+      },
+      len: {
+        args: [14, 14],
+        msg: "This field must be 11 characters! ex: 000.000.000-00",
+      },
+    },
   },
   birthday: {
     type: DataTypes.DATEONLY,
