+++ src/models/phone.model.js
@@ -1,0 +1,27 @@
+const { DataTypes } = require("sequelize");
+
+const ClientModel = require("../models/client.model");
+const db = require("../config/db.config");
+
+const Phone = db.define("phone", {
+  id: {
+    type: DataTypes.INTEGER,
+    autoIncrement: true,
+    allowNull: false,
+    primaryKey: true,
+  },
+  phone: {
+    type: DataTypes.STRING,
+  },
+});
+
+// client
+// 1:1
+Phone.belongsTo(ClientModel);
+
+// 1:M
+ClientModel.hasMany(Phone);
+
+Phone.sync(); // This creates the table if it doesn't exist (and does nothing if it already exists)
+
+module.exports = Phone;
